+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# [1.2.0-rc.5](https://github.com/janovix/aml-svc/compare/v1.2.0-rc.4...v1.2.0-rc.5) (2026-01-10)
+
+
+### Features
+
+* full sentry integration ([244b958](https://github.com/janovix/aml-svc/commit/244b958e31bc95d1121f630b683348fe4cafe211))
+
+# [1.2.0-rc.4](https://github.com/janovix/aml-svc/compare/v1.2.0-rc.3...v1.2.0-rc.4) (2026-01-09)
+
+
+### Features
+
+* add endpoint to cancel alerts with reason and implement corresponding schema and tests ([eb0ebe6](https://github.com/janovix/aml-svc/commit/eb0ebe67c4da50e2a98d07a222ad7741faa8943d))
+
+# [1.2.0-rc.3](https://github.com/janovix/aml-svc/compare/v1.2.0-rc.2...v1.2.0-rc.3) (2026-01-09)
+
+
+### Features
+
+* update client statistics retrieval to include physical and moral client counts ([01ca427](https://github.com/janovix/aml-svc/commit/01ca4274a5593cd830d3665447414fe78f547ecb))
+
+# [1.2.0-rc.2](https://github.com/janovix/aml-svc/compare/v1.2.0-rc.1...v1.2.0-rc.2) (2026-01-09)
 # [1.2.0](https://github.com/janovix/aml-svc/compare/v1.1.0...v1.2.0) (2026-01-09)
 
 
@@ -14,6 +36,7 @@
 
 ### Features
 
+* enhance search field validation in CatalogListQuerySchema to treat empty and whitespace-only strings as undefined ([178ae5e](https://github.com/janovix/aml-svc/commit/178ae5e8b1d415676d20a05a7571bc38d21db21d))
 * add Caddyfile for local development and update README with local dev instructions ([8f95e28](https://github.com/janovix/aml-svc/commit/8f95e281de5e5fc0ac087c5fee3719bd0047bdf7))
 * add endpoints for retrieving catalog items by ID and active alert rules for seekers ([ab78f97](https://github.com/janovix/aml-svc/commit/ab78f973f3e0a6c8efc3d632bf3e3f97a7e94a13))
 * add Notices API endpoints for managing SAT notices and their statuses ([e051bc6](https://github.com/janovix/aml-svc/commit/e051bc6801e48df5b9c01be15623c9ab10e74300))